refactor(server): lift flow components out of authflow to server level
The shared flow steps (mfa, consent, issue, logout) and their shared
infrastructure (session, web) are peer domains to grants, device and tokens,
not subordinate to the browser-login package. Move them from
server/authflow/* up to server/*, matching dex's flat domain layout. authflow
keeps only the login flow and the step-sequencing spine, and imports the
components as siblings.
