+++ server/authflow/mfa_test.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dexidp/dex/server/authflow/mfa"
-	"github.com/dexidp/dex/server/authflow/session"
 	"github.com/dexidp/dex/server/internal"
+	"github.com/dexidp/dex/server/mfa"
+	"github.com/dexidp/dex/server/session"
 	"github.com/dexidp/dex/storage"
 )
