+++ server/logout/doc.go
