+++ server/session/doc.go
